refactor(sequencer/profiling): Move pyroscope setup function to its own file
