+++ apps/sequencer/src/lib.rs
@@ -5,5 +5,6 @@ pub mod feeds;
 pub mod http_handlers;
 pub mod metrics_collector;
 pub mod providers;
+pub mod pyroscope_profiler;
 pub mod reporters;
 pub mod sequencer_state;
